feat: add unit tests for panel alignment and inset functionality
